cs/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
xgboost==2.0.3
shap==0.44.0
scikit-learn==1.4.0
imbalanced-learn==0.12.0
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
sqlalchemy==2.0.25
python-dotenv==1.0.0
pytest==8.0.0
